Se agrega llamado y transformación de la entidad platos y bebidas, tiempo de llamado a los servicios por configuración

+++ lab3servicesgr06/src/main/java/co/edu/udea/compumovil/gr06_20182/lab3/fragment/DrinkAddEdit.java
@@ -130,7 +130,7 @@ public class DrinkAddEdit extends Fragment {
                 drink.setName(txtNameDrink.getText().toString().trim());
                 drink.setFavorite(checkFavorite.isChecked());
                 drink.setImage(ImageHelper.imageViewToByte(imgDrink));
-                drink.setPrice(Integer.parseInt(txtPrice.getText().toString().trim()));
+                drink.setPrice(Float.parseFloat(txtPrice.getText().toString().trim()));
 
                 try {
                     if(isNew){
